pyproject.toml
[tool.poetry]
name = "shafigullinik_gendiff"
version = "0.3.2"
description = ""
authors = ["shafigullinik <shafigullin.ik@gmail.com>"]
packages = [{include = 'gendiff'},]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
